Move figure into seperate class
